refactor: streamline chat message fetching and layout components
- Removed unused imports and variables in chat-related components for cleaner code.
- Simplified message fetching logic by integrating decryption directly into the query function.
- Updated chat room list and item components to enhance performance and maintainability.
- Adjusted layout components to remove unnecessary styles and improve readability.
