+++ components/Chat/useMessageFetching.ts
@@ -3,11 +3,7 @@ import useAuth from '@/hooks/useAuth';
 import {
   getMessagesChatMessagesGetInfiniteOptions,
 } from '@/lib/api/generated/@tanstack/react-query.gen';
-import { getMessagesChatMessagesGet } from '@/lib/api/generated';
-import { ChatMessage, GetMessagesResponse } from '@/lib/api/generated';
-import ProtocolService from '@/lib/services/ProtocolService';
-
-type ChatMessages = ChatMessage[];
+import { decryptMessages } from '@/lib/utils';
 
 const useMessageFetching = (
   roomId: string,
@@ -23,84 +19,7 @@ const useMessageFetching = (
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage } =
     useInfiniteQuery({
       ...queryOptions,
-      queryFn: async ({ pageParam, queryKey, signal }) => {
-        const one = queryKey[0];
-
-        const { data } = await getMessagesChatMessagesGet({
-          query: {
-            page_size: one.query.page_size,
-            room_id: one.query.room_id,
-            page: pageParam as number,
-          },
-          signal,
-          throwOnError: true,
-        });
-
-        // Decrypt messages
-        const localUserId = user?.id;
-
-        if (!localUserId) {
-          return data;
-        }
-
-        let decryptedMessages: ChatMessages = [];
-        try {
-          const processedMessages = await Promise.all(
-            data.messages.map(async (message: ChatMessage) => {
-              try {
-                if (message.encrypted_content && message.nonce) {
-                  let decryptedMessage = '';
-
-                  decryptedMessage = await ProtocolService.decryptMessage(
-                    localUserId === message.author_id
-                      ? message.recipient_id
-                      : message.author_id,
-                    {
-                      encryptedMessage: message.encrypted_content,
-                      nonce: message.nonce,
-                    },
-                  );
-
-                  return {
-                    ...message,
-                    message: decryptedMessage,
-                  };
-                }
-                return message;
-              } catch (decryptError) {
-                console.error(
-                  `Failed to decrypt message ${message.id}:`,
-                  decryptError
-                );
-                return null; // Return null for failed messages
-              }
-            }),
-          );
-
-          // Filter out null values from failed decryption attempts
-          decryptedMessages = processedMessages.filter(
-            (msg): msg is ChatMessage => msg !== null,
-          );
-        } catch (error) {
-          console.error('Error processing messages', error);
-          decryptedMessages = data.messages
-            .map((message: ChatMessage) => {
-              if (message.encrypted_content) {
-                return null;
-              }
-              return message;
-            })
-            .filter((msg): msg is ChatMessage => msg !== null);
-        }
-
-        // Return the same structure as GetMessagesResponse but with decrypted messages
-        const response: GetMessagesResponse = {
-          ...data,
-          messages: decryptedMessages,
-        };
-
-        return response;
-      },
+      queryFn: decryptMessages(user),
       staleTime: Infinity,
       getNextPageParam: (lastPage, allPages) => {
         if (!lastPage.next_cursor) {
