+++ components/ChatItem/index.tsx
@@ -19,7 +19,6 @@ function ChatItem({ item }: { item: ChatRoom }) {
   const targetUser = item.participants.find(
     (user) => user.id !== authorizedUser.id,
   );
-
   const isNavigationEnabled = !!targetUser?.username;
 
   // Prefetch chat data when user interacts with chat item
@@ -101,7 +100,6 @@ function ChatItem({ item }: { item: ChatRoom }) {
     }
     return null;
   };
-
   return (
     <Link
       href={{
