+++ components/ChatRoomList/index.tsx
@@ -11,47 +11,48 @@ import { useQueryClient } from '@tanstack/react-query';
 import useUserChats from '@/hooks/useUserChats';
 import ChatItem from '../ChatItem';
 import useAuth from '@/hooks/useAuth';
+import { getMessageChatRoomOptions, getMessagesChatMessagesGetInfiniteOptions, getUserChatRoomsOptions, getUserChatRoomsQueryKey } from '@/lib/api/generated/@tanstack/react-query.gen';
+import { CHAT_PAGE_SIZE, decryptMessages } from '@/lib/utils';
+import { ChatRoom } from '@/lib/api/generated';
 
 export default function ChatRoomList() {
   const queryClient = useQueryClient();
   const { user } = useAuth();
+  const queryOptions = getUserChatRoomsOptions();
+
   const { chats, isFetching, refetch } = useUserChats({ poolMs: 5000 });
   const prefetchedChatsRef = useRef(new Set());
 
   const onRefresh = useCallback(() => {
     queryClient.invalidateQueries({
-      queryKey: ['user-chat-rooms'],
+      queryKey: queryOptions.queryKey,
     });
     refetch();
   }, [queryClient, refetch]);
 
   // Prefetch messages for each chat room when chat list is loaded
   useEffect(() => {
-    if (chats && chats.chat_rooms.length > 0) {
+    if (chats && chats.length > 0) {
       // Prefetch the first few chat rooms' messages
-      chats.chat_rooms.slice(0, 3).forEach((chat) => {
+      chats.slice(0, 3).forEach((chat) => {
         // Skip if already prefetched
         if (prefetchedChatsRef.current.has(chat.id)) {
           return;
         }
 
-        const recipientId =
-          chat.participants.find((p) => p.id !== user.id)?.id || '';
-
-        // // Prefetch message room data
-        // queryClient.prefetchQuery({
-        //   queryKey: ["user-chat-room", chat.id],
-        //   queryFn: () => api.getMessageRoom(chat.id),
-        //   staleTime: 60 * 1000, // 1 minute
-        // });
-
-        // // Prefetch first page of messages
-        // queryClient.prefetchInfiniteQuery({
-        //   queryKey: ["messages", chat.id],
-        //   queryFn: ({ pageParam = 1 }) =>
-        //     api.fetchMessages(pageParam, 30, chat.id, user.id),
-        //   initialPageParam: 1,
-        // });
+        const messageOptions = getMessagesChatMessagesGetInfiniteOptions({
+          query: {
+            page_size: CHAT_PAGE_SIZE,
+            room_id: chat.id,
+          },
+        });
+      
+        // Prefetch first page of messages
+        queryClient.prefetchInfiniteQuery({
+          ...messageOptions,
+          queryFn: decryptMessages(user),
+          initialPageParam: 1,
+        });
 
         // Mark as prefetched
         prefetchedChatsRef.current.add(chat.id);
@@ -60,8 +61,8 @@ export default function ChatRoomList() {
   }, [chats, queryClient, user.id]);
 
   function renderList() {
-    return chats?.chat_rooms.map((item) => (
-      <ChatItem key={item.id} item={item} />
+    return chats?.map((item) => (
+      <ChatItem key={item.id} item={item as ChatRoom} />
     ));
   }
 
@@ -77,7 +78,7 @@ export default function ChatRoomList() {
         ) : (
           <>
             {renderList()}
-            {!isFetching && !chats?.chat_rooms.length && (
+            {!isFetching && !chats?.length && (
               <View style={{ height: 100 }} />
             )}
           </>
