+++ hooks/useMessageRoom.ts
@@ -1,8 +1,5 @@
-import { getMessageChatRoom } from '@/lib/api/generated';
 import { getMessageChatRoomOptions } from '@/lib/api/generated/@tanstack/react-query.gen';
 import { useQuery } from '@tanstack/react-query';
-import { AxiosError } from 'axios';
-import { useEffect } from 'react';
 
 export default function useMessageRoom(
   roomId: string,
@@ -21,10 +18,8 @@ export default function useMessageRoom(
     }),
     retry: false,
     enabled,
-    staleTime: 1000 * 60, // Consider data fresh for 1 minute
-    gcTime: 1000 * 60 * 10, // Keep data in cache for 10 minutes
     refetchIntervalInBackground: false,
-    refetchOnWindowFocus: true,
+    refetchOnWindowFocus: false,
   });
 
   return { room, isFetching: isFetching && !isRefetching, error };
